Dopisanie testów dla AnimalService, dodanie listy todo

+++ src/main/java/pl/schronisko/demo/Animal/AnimalService.java
@@ -8,6 +8,7 @@ import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+
 @Service
 public class AnimalService {
 
@@ -25,48 +26,52 @@ public class AnimalService {
 
     public void addNewAnimal(Animal animal) {
         Optional<Animal> animalOptional = animalRepository.findAnimalByName(animal.getName());
-        if (animalOptional.isPresent()){
+        if (animalOptional.isPresent()) {
             throw new IllegalStateException("this name is already taken");
         }
         animalRepository.save(animal);
     }
 
     public void deleteAnimal(Long animalId) {
-          boolean exists =  animalRepository.existsById(animalId);
-                  if (!exists) {
-                      throw new IllegalStateException("animal with id " + animalId + " does not exists");
-                  }
-                  animalRepository.deleteById(animalId);
+        boolean exists = animalRepository.existsById(animalId);
+        if (animalId == null) {
+            throw new IllegalArgumentException("id cannot be null");
+        }
+        if (!exists) {
+            throw new IllegalStateException("animal with id " + animalId + " does not exists");
+        }
+        animalRepository.deleteById(animalId);
     }
 
     @Transactional
-    public void updateAnimal(Long animalId, String name, String category, Sex sex, Integer age, String color, Size size){
-        Animal animal = animalRepository.findById(animalId).orElseThrow(()->new IllegalStateException(
+    public void updateAnimal(Long animalId, String name, Category category, Sex sex, Integer age, String color, Size size) {
+        Animal animal = animalRepository.findById(animalId).orElseThrow(() -> new IllegalStateException(
                 "Animal with id " + animalId + " does not exists"));
-        if (name !=null && name.length()>0 && !Objects.equals(animal.getName(),name)
-        ){
+        if (name != null && name.length() > 0 && !Objects.equals(animal.getName(), name)
+        ) {
             animal.setName(name);
         }
-        if (category !=null && category.length()>0 && !Objects.equals(animal.getCategory(),category)
-        ){
+        if (category != null && !Objects.equals(animal.getCategory(), category)
+        ) {
             animal.setCategory(category);
         }
-        if (sex !=null  && !Objects.equals(animal.getSex(),sex)
-        ){
+        if (sex != null && !Objects.equals(animal.getSex(), sex)
+        ) {
             animal.setSex(sex);
         }
-        if (age !=null && age>0 && !Objects.equals(animal.getAge(),age)
-        ){
+        if (age != null && age > 0 && !Objects.equals(animal.getAge(), age)
+        ) {
             animal.setAge(age);
         }
-        if (color !=null && color.length()>0 && !Objects.equals(animal.getColor(),color)
-        ){
+        if (color != null && color.length() > 0 && !Objects.equals(animal.getColor(), color)
+        ) {
             animal.setColor(color);
         }
-        if (size !=null && !Objects.equals(animal.getSize(),size)
-        ){
+        if (size != null && !Objects.equals(animal.getSize(), size)
+        ) {
             animal.setSize(size);
         }
+        animalRepository.save(animal);
     }
 
 }
